pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentiment-span-extractor"
version = "0.1.0"
description = "Tweet Sentiment Span Extraction using PyTorch Lightning"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "MoiseevRoman", email = "rrmoiseev@edu.hse.ru"}
]
dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "transformers>=4.44.0",
    "tokenizers>=0.15.0",
    "hydra-core>=1.3.0",
    "mlflow>=2.0.0",
    "dvc>=3.0.0",
    "kaggle>=1.5.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "fire>=0.5.0",
    "rich>=13.0.0",
    "tritonclient[http]>=2.40.0",
    "onnx>=1.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pre-commit>=3.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
